Extracted turret ammo logic into a helper class (W.I.P).
